Moved MembersAPI out of index.js
refs https://github.com/TryGhost/Team/issues/879
